+++ ricaun.Revit.DA.Tests/Tests.cs
@@ -9,8 +9,10 @@ namespace ricaun.Revit.DA.Tests
 {
     public class Tests : BundleFileTests
     {
-        [TestCase("2024", "4.7", "19", null)]
-        [TestCase("2025", "4.7", "19", null)]
+        //[TestCase("2024", "4.7", "19", null)]
+        //[TestCase("2025", "4.7", "19", null)]
+        [TestCase("2024", "4.8", "21", "Example")]
+        [TestCase("2025", "8.0", "25", "Example")]
         public async Task ExecuteDesignAutomation(string engine, string frameworkNameContain, string referenceContain, string addInNameContain)
         {
             var bundlePaths = GetBundles();
@@ -22,12 +24,12 @@ namespace ricaun.Revit.DA.Tests
 
             Console.WriteLine(output.ToJson());
 
-            Assert.IsTrue(output.VersionName.Contains(engine), $"VersionName contains engine {engine}");
-            Assert.IsTrue(output.FrameworkName.Contains(frameworkNameContain), $"FrameworkName contains framework {frameworkNameContain}");
-            Assert.IsTrue(output.Reference.Contains(referenceContain), $"Reference contains reference {referenceContain}");
+            Assert.IsTrue(output.VersionName.Contains(engine), $"VersionName {output.VersionName} not contains engine {engine}");
+            Assert.IsTrue(output.FrameworkName.Contains(frameworkNameContain), $"FrameworkName {output.FrameworkName} not contains framework {frameworkNameContain}");
+            Assert.IsTrue(output.Reference.Contains(referenceContain), $"Reference {output.Reference} not contains reference {referenceContain}");
 
             if (output.AddInName is not null)
-                Assert.IsTrue(output.AddInName.Contains(addInNameContain), $"AddinName contains addin {addInNameContain}");
+                Assert.IsTrue(output.AddInName.Contains(addInNameContain), $"AddinName {output.AddInName} not contains addin {addInNameContain}");
             else
                 Assert.AreEqual(addInNameContain, output.AddInName);
         }